sched: Partially rework Shadow Scheduling for PA-18
This partially reworks Shadow Scheduling to match
the new ideas and improvements outlined in ticket
PA-18.

For this to happen we introduce a seperate method
__reevaluate_shadow_policy which decides which
scheduling policy should be chosen and whether
it is worth it to do a full task reclassification
instantly.

override_shadow indicates whether the regular policy
is more aggressive than the Shadow policy.
Based on that, we pick the more aggressive policy
and write it to proposed_upmigrate_pct and proposed_downmigrate_pct.

In the end the proposed policy gets compared against
our threshold SCHED_RECLASSIFY_THRESHOLD.

If it passes this last test, then we execute the necessary
conversions and write it to real_upmigrate/real_downmigrate
whilst executing a full task reclassification.

Change-Id: Iabb08e0c7663a728864ade0927ceb58ba32b7363
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: MOVZX <movzx@yahoo.com>