defconfig: Disable CONFIG_SCHED_FREQ_INPUT
As per Tim Murray @Google:

"I think we couldn't find any use cases where it helped and we hit some
other issue with HMP, plus we were so dependent on touch boost on that
device that it wasn't doing anything helpful with frequencies in the first
place."

And as shadow scheduling is on the way, we turned to relay on it instead
of this stuff.

This patch is inspired by xboxfanj.

Signed-off-by: Wang Han <416810799@qq.com>
Signed-off-by: MOVZX <movzx@yahoo.com>